+++ bindings/nostr-sdk-ffi/src/protocol/event/tag/mod.rs
@@ -203,6 +203,8 @@ impl Tag {
     }
 
     /// Compose `["t", "<hashtag>"]` tag
+    ///
+    /// This will convert the hashtag to lowercase.
     #[uniffi::constructor]
     pub fn hashtag(hashtag: &str) -> Self {
         Self {